Add files via upload
Updated Galerkin1DElliptic so we can use the conductive field as a input vector rather than an input lambda function

The NISP is the goal so far. I added Hermite polynomial to get equations such as (2.45) in the Spectral Methods for UQ book. This functions returns a vector of dimension P+1 with entries $\Psi_k$ in the same order as in the book. 

The NISP is not working and I don't know why...